feat(db): initialize and close DB with app context  Added app in the call to initialize the database for the app in app/init.py. Included app in the initialize DB function with app context and teardown to close the database when the server shuts down. Added close_db function to db/init.py. Added get_db function to routes/home.py to query posts created at a certain time.  #3 Pass User Data into Homepage Templates
feat(db): initialize and close DB with app context

Added app in the call to initialize the database for the app in app/init.py. Included app in the initialize DB function with app context and teardown to close the database when the server shuts down. Added close_db function to db/init.py. Added get_db function to routes/home.py to query posts created at a certain time.

#3 Pass User Data into Homepage Templates